common/mlx5: Add common primary/secondary processes handling codes
Move common mp handling codes from net pmd to common drive.
